Refactor redux structure (#9)
* Refactor Actions to separate files ♻

* Update useSubscribe to enable custom Action Creators

* Change gameReducer to readyReducer. Add updatePlayersReady

* Refactor folder structure reducers
