+++ front-end/src/actions/readyActions.js
@@ -1,0 +1,9 @@
+import { PLAYER_READY, UPDATE_PLAYERS_READY } from './types';
+
+export const toggleReady = () => {
+  return { type: PLAYER_READY };
+};
+
+export const updatePlayersReady = (topic, playersReady) => {
+  return { type: UPDATE_PLAYERS_READY, payload: playersReady };
+};
