+++ front-end/src/actions/types.js
@@ -2,7 +2,7 @@
 export const GET_DATA = 'GET_DATA';
 export const POST_DATA = 'POST_DATA';
 
-// Forms
+// --------------- Forms -----------------------
 export const FORM_SUBMIT = 'FORM_SUBMIT';
 export const FORM_VALID = 'FORM_VALID';
 export const FORM_ERROR = 'FORM_ERROR';
@@ -12,7 +12,9 @@ export const RESET_FORM = 'RESET_FORM';
 export const SET_CLIENT = 'SET_CLIENT';
 export const SET_CONNECTED = 'SET_CONNECTED';
 
-// Subscriptions
+// -------------- Subscriptions -------------------------
 export const SUBSCRIPTION_MESSAGE = 'SUBSCRIPTION_MESSAGE';
 
-export const READY = 'READY';
+// --------------------- Ready -------------------------
+export const PLAYER_READY = 'READY';
+export const UPDATE_PLAYERS_READY = 'UPDATE_PLAYERS_READY';
