+++ front-end/src/components/testing/GreetButton.js
@@ -3,9 +3,10 @@ import usePublish from '../../api/usePublish';
 import useSubscribe from '../../api/useSubscribe';
 import { useSelector } from 'react-redux';
 import { Link } from 'react-router-dom';
+import { addSubMessage } from '../../actions';
 
 export default (props) => {
-  useSubscribe('/topic/greeting');
+  useSubscribe('/topic/greeting', addSubMessage);
   const message = useSelector((state) => state.subscription['/topic/greeting']);
   const greetPublish = usePublish({
     destination: '/app/greeting',
