+++ front-end/src/components/testing/ReadyUpButton.js
@@ -2,19 +2,19 @@ import React, { useEffect } from 'react';
 import usePublish from '../../api/usePublish';
 import useSubscribe from '../../api/useSubscribe';
 import { useSelector, useDispatch } from 'react-redux';
-import { toggleReady } from '../../actions';
+import { toggleReady, updatePlayersReady } from '../../actions';
 
 export default (props) => {
   const dispatch = useDispatch();
   const subDest = '/topic/ready';
-  const playersReady = useSelector((state) => state.subscription[subDest]);
-  const ready = useSelector((state) => state.game.ready);
+  const playersReady = useSelector((state) => state.ready.totalReady);
+  const ready = useSelector((state) => state.ready.playerReady);
   const publishToggleReady = usePublish({
     destination: '/app/ready',
     body: ready.toString(), // Can be optimized
   });
 
-  useSubscribe(subDest);
+  useSubscribe(subDest, updatePlayersReady);
 
   // const buttonMsg = ready ? 'Not Ready' : 'Ready';
   const buttonClassName = ready ? 'active' : 'inactive';
