--- front-end/src/reducers/gameReducer.js
@@ -1,9 +1,0 @@
-export default (state = { ready: false }, { type, payload }) => {
-  switch (type) {
-    case 'READY':
-      const ready = !state.ready;
-      return { ...state, ready };
-    default:
-      return state;
-  }
-};
