+++ front-end/src/reducers/index.js
@@ -1,15 +1,15 @@
 import { combineReducers } from 'redux';
-import restReducer from './restReducer';
-import { clientReducer, connectedClientReducer } from './clientReducer';
-import subReducer from './subReducer';
-import gameReducer from './gameReducer';
+import restReducer from './API/restReducer';
+import { clientReducer, connectedClientReducer } from './API/clientReducer';
+import subReducer from './API/subReducer';
 import joinFormReducer from './joinFormReducer';
+import readyReducer from './readyReducer';
 
 export default combineReducers({
   data: restReducer,
   client: clientReducer,
   connected: connectedClientReducer,
   subscription: subReducer,
-  game: gameReducer,
+  ready: readyReducer,
   form: joinFormReducer,
 });
